requirements.txt
# application
flask
python-dotenv

# tests
pytest
pytest-watch
requests

# operations
gunicorn
